research/plugin_sources/nonebot-plugin-bili-helper-master/pyproject.toml
[tool.poetry]
name = "nonebot_plugin_bili_helper"
version = "0.9.3"
description = "Nonebot2 plugin for bilibili sharing detail"
authors = ["krimeshu <krimeshu@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/krimeshu/nonebot-plugin-bili-helper"
repository = "https://github.com/krimeshu/nonebot-plugin-bili-helper"

[tool.poetry.dependencies]
aiohttp = "^3.12.15"
python = "^3.9"
nonebot2 = "^2.3.0"
nonebot-adapter-onebot = "^2.0.0"
nonebot-plugin-htmlrender = "^0.6.0"
nonebot-plugin-localstore = "^0.7.4"

[tool.poetry.group.dev.dependencies]
nonebot2 = { extras = ["fastapi", "httpx"], version = "^2.3.0" }
nonebot-adapter-onebot = "^2.2.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = []

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
